Transactions table integrated and seed file completed
