+++ backend/models/accounts.py
@@ -15,4 +15,4 @@ class Account(db.Model, SerializerMixin):
     user = db.relationship('User', back_populates='accounts')
     transactions = db.relationship('Transaction', back_populates='account')
 
-
+from models.transactions import Transaction
