+++ backend/models/users.py
@@ -4,6 +4,8 @@ from sqlalchemy.orm import validates
 from sqlalchemy.ext.hybrid import hybrid_property
 from sqlalchemy_serializer import SerializerMixin
 
+from models.accounts import Account
+
 
 
 class User(db.Model, SerializerMixin):
